+++ changes.md
@@ -7,7 +7,8 @@ Here **[Breaking]** means the change that may break valid user code, **[Fix]** -
 
 * **[Breaking]** ampls::CPLEXCallback is now using the generic callbacks. Note that multithreading 
   is not disabled by default but implementation needs extra care. 
-  
+* **[Breaking]** Solver and solver driver related errors in Python are now thrown as 
+  ampls.AMPLSolverException
 
 ## 0.1.2 - 2023-06-02